Call stack tree

Shows the relationship between calls on the stack. This function combines the results of sys.calls() and sys.parents() yielding a display that shows how frames on the call stack are related.

cst()

Examples

# If all evaluation is eager, you get a single tree
f <- function() g()
g <- function() h()
h <- function() cst()
f()

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             └─f()
#>                               └─g()
#>                                 └─h()
#>                                   └─cst()

# You get multiple trees with delayed evaluation
try(f())

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             ├─try(f())
#>                             │ └─tryCatch(...)
#>                             │   └─tryCatchList(expr, classes, parentenv, handlers)
#>                             │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                             │       └─doTryCatch(return(expr), name, parentenv, handler)
#>                             └─f()
#>                               └─g()
#>                                 └─h()
#>                                   └─cst()

# Pay attention to the first element of each subtree: each
# evaluates the outermost call
f <- function(x) g(x)
g <- function(x) h(x)
h <- function(x) x
try(f(cst()))

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             ├─try(f(cst()))
#>                             │ └─tryCatch(...)
#>                             │   └─tryCatchList(expr, classes, parentenv, handlers)
#>                             │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                             │       └─doTryCatch(return(expr), name, parentenv, handler)
#>                             ├─f(cst())
#>                             │ └─g(x)
#>                             │   └─h(x)
#>                             └─cst()

# With a little ingenuity you can use it to see how NSE
# functions work in base R
with(mtcars, {cst(); invisible()})

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             ├─with(...)
#>                             └─with.default(...)
#>                               └─eval(substitute(expr), data, enclos = parent.frame())
#>                                 └─eval(substitute(expr), data, enclos = parent.frame())
#>                                   └─cst()

invisible(subset(mtcars, {cst(); cyl == 0}))

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             ├─subset(...)
#>                             └─subset.data.frame(...)
#>                               └─eval(e, x, parent.frame())
#>                                 └─eval(e, x, parent.frame())
#>                                   └─cst()

# You can also get unusual trees by evaluating in frames
# higher up the call stack
f <- function() g()
g <- function() h()
h <- function() eval(quote(cst()), parent.frame(2))
f()

#> █
#> ├─tryCatch(...)
#> │ └─tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─tryCatchOne(...)
#> │   │ └─doTryCatch(return(expr), name, parentenv, handler)
#> │   └─tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─doTryCatch(return(expr), name, parentenv, handler)
#> ├─withCallingHandlers(...)
#> ├─saveRDS(...)
#> ├─do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─build_site_local(...)
#>       └─build_reference(...)
#>         └─purrr::map(...)
#>           └─.f(.x[[i]], ...)
#>             └─data_reference_topic(...)
#>               ├─as_data(...)
#>               └─as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate_call(...)
#>                         ├─timing_fn(...)
#>                         ├─handle(...)
#>                         │ └─try(f, silent = TRUE)
#>                         │   └─tryCatch(...)
#>                         │     └─tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─withCallingHandlers(...)
#>                         ├─withVisible(eval(expr, envir, enclos))
#>                         └─eval(expr, envir, enclos)
#>                           └─eval(expr, envir, enclos)
#>                             └─f()
#>                               ├─g()
#>                               │ └─h()
#>                               │   └─eval(quote(cst()), parent.frame(2))
#>                               │     └─eval(quote(cst()), parent.frame(2))
#>                               └─cst()