Inspect an object

obj_inspect(x) is similar to .Internal(inspect(x)). The main difference is the output is a little more compact, it recurses fully, and avoids getting stuck in infinite loops by using a depth-first search. It also returns a list that you can compute with, and carefully uses colour to highlight the most important details.

obj_inspect(x, expand = character())

Arguments

x
Object to inspect

expand
Optionally, expand components of the true that are usually suppressed. Use:

"character" to show underlying entries in the global string pool.

"environment" to show the underlying hashtables.

"altrep" to show the underlying data.

"call" to show the full AST (but ast() is usually superior)

"bytecode" to show generated bytecode.

Examples

x <- list(
  TRUE,
  1L,
  runif(100),
  "3"
)
obj_inspect(x)

#> [1:0x7114ea8] <VECSXP[4]> (named:7)
#>   [2:0x3b1a928] <LGLSXP[1]> (named:7)
#>   [3:0x3b1a960] <INTSXP[1]> (named:7)
#>   [4:0x441b5e0] <REALSXP[100]> (named:0)
#>   [5:0x3b1aa08] <STRSXP[1]> (named:7)

# Expand "character" to see underlying CHARSXP entries in the global
# string pool
x <- c("banana", "banana", "apple", "banana")
obj_inspect(x)

#> [1:0x8bb0328] <STRSXP[4]> (named:7)

obj_inspect(x, expand = "character")

#> [1:0x8bb0328] <STRSXP[4]> (named:7)
#>   [2:0x3b1adc0] <CHARSXP> (named:0)
#>   [2:0x3b1adc0]
#>   [3:0x3344500] <CHARSXP> (named:0)
#>   [2:0x3b1adc0]

# Expand altrep to see underlying data
x <- 1:10
obj_inspect(x)

#> [1:0x8c190d8] <INTSXP[10]> (altrep named:7)

obj_inspect(x, expand = "altrep")

#> [1:0x8c190d8] <INTSXP[10]> (altrep named:7)
#>   _class [2:0x228e290] <RAWSXP[144]> (named:0)
#>     _attrib [3:0x22c4e98] <LISTSXP> (named:0)
#>       [4:0x22bd0e8] <SYMSXP: compact_intseq> (named:0)
#>       [5:0x228fa70] <SYMSXP: base> (named:7)
#>       [6:0x22c28b0] <INTSXP[1]> (named:0)
#>   _data1 [7:0x8d52e18] <REALSXP[3]> (named:0)
#>   _data2 <NILSXP>

# Expand environmnets to see the underlying implementation details
e1 <- new.env(hash = FALSE, parent = emptyenv(), size = 3L)
e2 <- new.env(hash = TRUE, parent = emptyenv(), size = 3L)
e1$x <- e2$x <- 1:10

obj_inspect(e1)

#> [1:0x5de71a0] <ENVSXP> (named:7)
#>   x [2:0x60a6788] <INTSXP[10]> (altrep named:7)
#>   _enclos [3:0x2290c98] <ENVSXP: empty> (named:7)

obj_inspect(e1, expand = "environment")

#> [1:0x5de71a0] <ENVSXP> (named:7)
#>   _frame [2:0x60a6e50] <LISTSXP> (named:0)
#>     x [3:0x60a6788] <INTSXP[10]> (altrep named:7)
#>   _hashtab <NILSXP>
#>   _enclos [5:0x2290c98] <ENVSXP: empty> (named:7)

obj_inspect(e2, expand = "environment")

#> [1:0x5f58e70] <ENVSXP> (named:7)
#>   _frame <NILSXP>
#>   _hashtab [3:0x7de0498] <VECSXP[3/1]> (named:0)
#>     [4:0x60a6b40] <LISTSXP> (named:0)
#>       x [5:0x60a6788] <INTSXP[10]> (altrep named:7)
#>     <NILSXP>
#>     <NILSXP>
#>   _enclos [6:0x2290c98] <ENVSXP: empty> (named:7)