Display tree of references

This tree display focusses on the distinction between names and values. For each reference-type object (lists, environments, and optional character vectors), it displays the location of each component. The display shows the connection between shared references using a locally unique id.

ref(..., character = FALSE)

Arguments

...

One or more objects

character

If TRUE, show references from character vector in to global string pool

See also

Examples

x <- 1:100
ref(x)
#> [1:0x5648ed07a010] <int> 

y <- list(x, x, x)
ref(y)
#> █ [1:0x5648ef117d68] <list> 
#> ├─[2:0x5648ed07a010] <int> 
#> ├─[2:0x5648ed07a010] 
#> └─[2:0x5648ed07a010] 
ref(x, y)
#> [1:0x5648ed07a010] <int> 
#>  
#> █ [2:0x5648ef117d68] <list> 
#> ├─[1:0x5648ed07a010] 
#> ├─[1:0x5648ed07a010] 
#> └─[1:0x5648ed07a010] 

e <- new.env()
e$e <- e
e$x <- x
e$y <- list(x, e)
ref(e)
#> █ [1:0x5648ec0a0070] <env> 
#> ├─x = [2:0x5648ed07a010] <int> 
#> ├─y = █ [3:0x5648ef0c5a08] <list> 
#> │     ├─[2:0x5648ed07a010] 
#> │     └─[1:0x5648ec0a0070] 
#> └─e = [1:0x5648ec0a0070] 

# Can also show references to global string pool if requested
ref(c("x", "x", "y"))
#> [1:0x5648ef49db78] <chr> 
ref(c("x", "x", "y"), character = TRUE)
#> █ [1:0x5648ef4745d8] <chr> 
#> ├─[2:0x5648e98ac700] <string: "x"> 
#> ├─[2:0x5648e98ac700] 
#> └─[3:0x5648e99a49e0] <string: "y">