tree_label {lobstr}	R Documentation

Build element or node label in tree

Description

These methods control how the value of a given node is printed. New methods can be added if support is needed for a novel class

Usage

tree_label(x, opts)

Arguments

x
A tree like object (list, etc.)

opts
A list of options that directly mirrors the named arguments of tree. E.g. list(val_printer = crayon::red) is equivalent to tree(..., val_printer = crayon::red).

[Package lobstr version 1.1.2 Index]