Call stack tree

Shows the relationship between calls on the stack. This function combines the results of sys.calls() and sys.parents() yielding a display that shows how frames on the call stack are related.

cst()

Examples

# If all evaluation is eager, you get a single tree
f <- function() g()
g <- function() h()
h <- function() cst()
f()

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             └─global::f()
#>                               └─global::g()
#>                                 └─global::h()
#>                                   └─lobstr::cst()

# You get multiple trees with delayed evaluation
try(f())

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             ├─base::try(f())
#>                             │ └─base::tryCatch(...)
#>                             │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                             │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                             │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                             └─global::f()
#>                               └─global::g()
#>                                 └─global::h()
#>                                   └─lobstr::cst()

# Pay attention to the first element of each subtree: each
# evaluates the outermost call
f <- function(x) g(x)
g <- function(x) h(x)
h <- function(x) x
try(f(cst()))

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             ├─base::try(f(cst()))
#>                             │ └─base::tryCatch(...)
#>                             │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                             │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                             │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                             ├─global::f(cst())
#>                             │ └─global::g(x)
#>                             │   └─global::h(x)
#>                             └─lobstr::cst()

# With a little ingenuity you can use it to see how NSE
# functions work in base R
with(mtcars, {cst(); invisible()})

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             ├─base::with(...)
#>                             └─base::with.default(...)
#>                               └─base::eval(substitute(expr), data, enclos = parent.frame())
#>                                 └─base::eval(substitute(expr), data, enclos = parent.frame())
#>                                   └─lobstr::cst()

invisible(subset(mtcars, {cst(); cyl == 0}))

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             ├─base::subset(...)
#>                             └─base::subset.data.frame(...)
#>                               └─base::eval(e, x, parent.frame())
#>                                 └─base::eval(e, x, parent.frame())
#>                                   └─lobstr::cst()

# You can also get unusual trees by evaluating in frames
# higher up the call stack
f <- function() g()
g <- function() h()
h <- function() eval(quote(cst()), parent.frame(2))
f()

#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     ├─withr::with_options(...)
#>                     │ └─base::force(code)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             └─global::f()
#>                               ├─global::g()
#>                               │ └─global::h()
#>                               │   └─base::eval(quote(cst()), parent.frame(2))
#>                               │     └─base::eval(quote(cst()), parent.frame(2))
#>                               └─lobstr::cst()