How much memory is currently used by R?

mem_used() wraps around gc() and returns the exact number of bytes currently used by R. Note that changes will not match up exactly to obj_size() as session specific state (e.g. .Last.value) adds minor variations.

Usage

mem_used()

Examples

prev_m <- 0; m <- mem_used(); m - prev_m
#> 61.36 MB

x <- 1:1e6
prev_m <- m; m <- mem_used(); m - prev_m
#> 86.32 kB
obj_size(x)
#> 680 B

rm(x)
prev_m <- m; m <- mem_used(); m - prev_m
#> 46.67 kB

prev_m <- m; m <- mem_used(); m - prev_m
#> 464 B