Display tree of references

This tree display focusses on the distinction between names and values. For each reference-type object (lists, environments, and optional character vectors), it displays the location of each component. The display shows the connection between shared references using a locally unique id.

ref(..., character = FALSE)

Arguments

...
One or more objects

character
If TRUE, show references from character vector in to global string pool

Examples

x <- 1:100
ref(x)

#> [1:0x64af590] <int> 

y <- list(x, x, x)
ref(y)

#> █ [1:0x6b30a08] <list> 
#> ├─[2:0x64af590] <int> 
#> ├─[2:0x64af590] 
#> └─[2:0x64af590] 

ref(x, y)

#> [1:0x64af590] <int> 
#>  
#> █ [2:0x6b30a08] <list> 
#> ├─[1:0x64af590] 
#> ├─[1:0x64af590] 
#> └─[1:0x64af590] 

e <- new.env()
e$e <- e
e$x <- x
e$y <- list(x, e)
ref(e)

#> █ [1:0x88cf320] <env> 
#> ├─x = [2:0x64af590] <int> 
#> ├─y = █ [3:0x8b01478] <list> 
#> │     ├─[2:0x64af590] 
#> │     └─[1:0x88cf320] 
#> └─e = [1:0x88cf320] 

# Can also show references to global string pool if requested
ref(c("x", "x", "y"))

#> [1:0x8df6988] <chr> 

ref(c("x", "x", "y"), character = TRUE)

#> █ [1:0x5c79ae8] <chr> 
#> ├─[2:0x25692a8] <string: "x"> 
#> ├─[2:0x25692a8] 
#> └─[3:0x26710f8] <string: "y">